change runtime datatype to double

+++ src/result.cpp
@@ -54,7 +54,7 @@ void writefile(vector<string> kartu, vector<string> *hasil, double runtime){
     }
 }
 
-void printruntime(float runtime){
+void printruntime(double runtime){
     cout << "============================" << endl;
     cout << "| Waktu Eksekusi : " << runtime << " ms    |" << endl;
     cout << "============================" << endl;
